+++ modules/apps/apio-architect/apio-architect-impl/src/test/java/com/liferay/apio/architect/internal/message/json/hal/HALMessageMapperTest.java
@@ -59,6 +59,8 @@ public class HALMessageMapperTest implements RepresentableManager {
 			Paths.get("src", "test", "resources", "hal")
 		).mediaType(
 			"application/hal+json"
+		).validateErrorMessageMapper(
+			new HALErrorMessageMapper()
 		).validateEntryPointMessageMapper(
 			new HALEntryPointMessageMapper()
 		).validatePageMessageMapper(