+++ modules/apps/asset/asset-publisher-web/src/main/java/com/liferay/asset/publisher/web/internal/portlet/action/AddAssetListMVCActionCommand.java
@@ -15,8 +15,6 @@
 package com.liferay.asset.publisher.web.internal.portlet.action;
 
 import com.liferay.asset.kernel.model.AssetEntry;
-import com.liferay.asset.list.constants.AssetListEntryTypeConstants;
-import com.liferay.asset.list.model.AssetListEntry;
 import com.liferay.asset.list.service.AssetListEntryService;
 import com.liferay.asset.publisher.constants.AssetPublisherPortletKeys;
 import com.liferay.asset.publisher.util.AssetPublisherHelper;
@@ -32,9 +30,14 @@ import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextFactory;
 import com.liferay.portal.kernel.servlet.MultiSessionMessages;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.UnicodeProperties;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
 
+import java.util.Enumeration;
 import java.util.List;
 import java.util.Objects;
 
@@ -81,7 +84,10 @@ public class AddAssetListMVCActionCommand extends BaseMVCActionCommand {
 			"selectionStyle", "dynamic");
 
 		try {
-			if (Objects.equals(selectionStyle, "manual")) {
+			if (Objects.equals(selectionStyle, "dynamic")) {
+				_saveDynamicAssetList(actionRequest, title, portletPreferences);
+			}
+			else if (Objects.equals(selectionStyle, "manual")) {
 				_saveManualAssetList(actionRequest, title, portletPreferences);
 			}
 
@@ -105,6 +111,39 @@ public class AddAssetListMVCActionCommand extends BaseMVCActionCommand {
 		}
 	}
 
+	private void _saveDynamicAssetList(
+			ActionRequest actionRequest, String title,
+			PortletPreferences portletPreferences)
+		throws Exception {
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		ServiceContext serviceContext = ServiceContextFactory.getInstance(
+			actionRequest);
+
+		UnicodeProperties properties = new UnicodeProperties(true);
+
+		Enumeration<String> names = portletPreferences.getNames();
+
+		while (names.hasMoreElements()) {
+			String name = names.nextElement();
+
+			String value = StringUtil.merge(
+				portletPreferences.getValues(name, null));
+
+			if (Validator.isNull(value)) {
+				continue;
+			}
+
+			properties.put(name, value);
+		}
+
+		_assetListEntryService.addDynamicAssetListEntry(
+			themeDisplay.getUserId(), themeDisplay.getScopeGroupId(), title,
+			properties.toString(), serviceContext);
+	}
+
 	private void _saveManualAssetList(
 			ActionRequest actionRequest, String title,
 			PortletPreferences portletPreferences)
@@ -116,11 +155,6 @@ public class AddAssetListMVCActionCommand extends BaseMVCActionCommand {
 		ServiceContext serviceContext = ServiceContextFactory.getInstance(
 			actionRequest);
 
-		AssetListEntry assetListEntry =
-			_assetListEntryService.addAssetListEntry(
-				themeDisplay.getScopeGroupId(), title,
-				AssetListEntryTypeConstants.TYPE_MANUAL, serviceContext);
-
 		long[] groupIds = _assetPublisherHelper.getGroupIds(
 			portletPreferences, themeDisplay.getScopeGroupId(),
 			themeDisplay.getLayout());
@@ -129,10 +163,12 @@ public class AddAssetListMVCActionCommand extends BaseMVCActionCommand {
 			actionRequest, portletPreferences,
 			themeDisplay.getPermissionChecker(), groupIds, true, true);
 
-		for (AssetEntry assetEntry : assetEntries) {
-			_assetListEntryService.addAssetEntrySelection(
-				assetListEntry.getAssetListEntryId(), assetEntry.getEntryId());
-		}
+		long[] assetEntryIds = ListUtil.toLongArray(
+			assetEntries, AssetEntry::getEntryId);
+
+		_assetListEntryService.addManualAssetListEntry(
+			themeDisplay.getUserId(), themeDisplay.getScopeGroupId(), title,
+			assetEntryIds, serviceContext);
 	}
 
 	@Reference